ddl: refine doc and test, and move test to a new file
