+++ pkg/ddl/tests/partition/BUILD.bazel
@@ -9,6 +9,7 @@ go_test(
         "exchange_partition_test.go",
         "global_index_version_test.go",
         "main_test.go",
+        "modify_column_test.go",
         "multi_domain_test.go",
         "placement_test.go",
         "reorg_partition_test.go",
